Improve cancelability of most games.
"from the I-wish-there-was-a-better-way-to-do-this dept."

Scatter checks for cancellation through new_game_desc in all the puzzles with wild abandon. There are probably too many. I have only vaguely checked that this is cancelable enough, and not checked at all whether I've inserted any checks in pointless places. This could really do with reviewing by people who know how the puzzles work.

The effect of these checks is that if the UI has set the flag meaning "user has pressed Cancel", then bomb out of new_game_desc as quickly and tidily as possible, having it return NULL.

The UI thread currently blocks on the cancel operation with Thread.join(), so the risk in the case of insufficient cancelability is one of "Application not responding for 5 seconds, [press here to kill -9]" rather than disastrous concurrency problems.