Fix #147: rotate presets to current orientation.
We lie a little and just tick the preset that matches modulo orientation; I think this is fine? (Responsibility for ticking it has moved up into Java.)

Also in terms of compile flags to vary available presets and game layout (reduce borders), declare that we are "small screen" and not "slow system", now that we have that option.

Change progress message to "Generating game" to reinforce that it is making one up, instead of just saying something vague about slowness.