Fix #188: ignore chooser choice within 300ms of open.
This seems short enough that it's very unlikely anyone has deliberately identified and clicked on a puzzle they wanted within that time, and much more likely that it's an accidental click.

That scenario will be further mitigated by #191, but this still seems worth having as well.

The ignoring is done with a touch listener so you don't get a background flash, nor a click sound effect, where applicable.