Fix #198: change behaviour of "Unlocked" to use "unspecified".
This means if the OS has locked orientation, we don't set it back to sensor, which would be inconsistent with the chooser and generally unexpected.

No need for migration code since old values of "sensor" will be caught by the default case anyway, and I don't think it's worth offering "sensor" in its own right.