Improved USB/Bluetooth(?) mouse support
This improves on the support for multi-button mice, at least on my
phone.  When using a USB mouse via a USB OTG adapter on my Galaxy S7,
the right mouse button sends a "back" KeyEvent rather than a standard
touch MotionEvent, with event.getSource() set to SOURCE_MOUSE.  So, we
track the mouse cursor location with onGenericMotionEvent(), and when we
get a Back press from the mouse, we dispatch it as a right button press
and/or press-and-hold.

One complication is that Android (at least on my GS7) will translate a
long-pressed Back button (even when delivered by a right mouse button)
into a Menu button, which will both pop up the in-game menu and also
implicitly release the "Back" button.  This behavior can only be
disabled at the Activity level, in dispatchKeyEvent(), so GamePlay now
delegates that method to GameView for KEYCODE_MENU events so that it can
be ignored.

Since we now have a working right-button implementation, we can now also
eliminate the touch-and-hold delay for primary click events when they
come from a mouse.

The Internet suggests that Bluetooth mice function identically to USB
mice, so those should have improved support as well.  It also suggests
that the right-click as Back button behavior may be
manufacturer-dependent, so this code might not work the same on all
Android devices.  However, the existing multibutton support should still
work, and those devices should also be able to take advantage of the
non-delayed click functionality.